+++ lib/graphql/fragment_cache.rb
@@ -21,6 +21,7 @@ module GraphQL
   module FragmentCache
     class << self
       attr_reader :cache_store
+      attr_accessor :namespace
 
       def use(schema_defn, options = {})
         verify_interpreter_and_analysis!(schema_defn)
@@ -30,6 +31,10 @@ module GraphQL
         schema_defn.extend(Schema::Patch)
 
         GraphQL::Pagination::Connections.prepend(Connections::Patch)
+      end
+
+      def configure
+        yield self
       end
 
       def cache_store=(store)