+++ spec/graphql/fragment_cache/field_extensions_spec.rb
@@ -272,8 +272,56 @@ describe "cache_fragment: option" do
     end
   end
 
-  context "when :force_cache is in the context" do
-    let(:context) { {force_cache: true} }
+  context "when :if is evaluating to true" do
+    let(:context) { {current_user: nil} }
+    let(:cache_fragment) { {if: -> { context[:current_user].nil? }} }
+
+    it "uses the cache" do
+      expect(execute_query.dig("data", "post")).to eq({
+        "id" => "1",
+        "title" => "option test"
+      })
+    end
+  end
+
+  context "when :if is evaluating to false" do
+    let(:context) { {current_user: User.new(id: "1", name: "some-user")} }
+    let(:cache_fragment) { {if: -> { context[:current_user].nil? }} }
+
+    it "does not use the cache" do
+      expect(execute_query.dig("data", "post")).to eq({
+        "id" => "1",
+        "title" => "new option test"
+      })
+    end
+  end
+
+  context "when :unless is evaluating to true" do
+    let(:context) { {current_user: nil} }
+    let(:cache_fragment) { {unless: -> { context[:current_user].nil? }} }
+
+    it "does not use the cache" do
+      expect(execute_query.dig("data", "post")).to eq({
+        "id" => "1",
+        "title" => "new option test"
+      })
+    end
+  end
+
+  context "when :unless is evaluating to false" do
+    let(:context) { {current_user: User.new(id: "1", name: "some-user")} }
+    let(:cache_fragment) { {unless: -> { context[:current_user].nil? }} }
+
+    it "uses the cache" do
+      expect(execute_query.dig("data", "post")).to eq({
+        "id" => "1",
+        "title" => "option test"
+      })
+    end
+  end
+
+  context "when :renew_cache is in the context" do
+    let(:context) { {renew_cache: true} }
 
     it "forces a cache miss and stores the computed value in the cache" do
       expect(execute_query.dig("data", "post")).to eq({
@@ -284,7 +332,7 @@ describe "cache_fragment: option" do
       # make object dirty
       post.title = "new option test 2"
 
-      context[:force_cache] = false
+      context[:renew_cache] = false
       expect(execute_query.dig("data", "post")).to eq({
         "id" => "1",
         "title" => "new option test"